feat: refactor WebSocket handling to utilize Dsn path method and streamline endpoint management
