+++ nodejs/test/bulkPulling/wsClient.recovery.test.ts
@@ -1,8 +1,7 @@
 import { w3cwebsocket } from "websocket";
 import { WsClient } from "../../src/client/wsClient";
 import { WebSocketConnectionPool } from "../../src/client/wsConnectorPool";
-import { parse } from "../../src/common/dsn";
-import { WS_TMQ_ENDPOINT } from "../../src/common/constant";
+import { parse, WS_TMQ_ENDPOINT } from "../../src/common/dsn";
 
 function createMockConnector() {
     return {
